requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
Django==4.1.3
django-environ==0.9.0
psycopg2
djangorestframework==3.12.2
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
dnspython==2.2.1
Pillow==9.3.0
gunicorn==20.1.0
pytz==2022.6
sqlparse==0.2.4
tzdata==2022.6
drf-spectacular